fix(markdown): unescape pipes in table-cell code spans
GFM strips a backslash-escaped pipe when it splits a table row into
cells, including inside other inline spans; leaving it to inline
parsing missed code spans, where a backslash is not an escape. Strip
at cell-split time with an index-to-source-offset map, and let the
CST own the consumed escape as an Escape node punched into code-span
interiors, so the direct parse and the lowered CST agree and tiling
holds. GFM conformance reaches its full measured set, 663/663;
CommonMark stays 652/652. Releases intent 0035 with its capability
link.

Also lands the GFM review follow-ups:

- tests: kyo-ui delete and task-list checkbox markup, genuinely
  empty table cells, MdStyle table knobs off their defaults, an
  assertion behind the conformance shared-base report line, and
  multi-line task-item marker ordering
- structure: table rows extracted to TableRows.scala, extended
  autolinks to ExtendedAutolinks.scala, conformance test scaffolding
  deduplicated into ConformanceFixtures.scala, Scalatags table cases
  hand-build trees per the file's own scaladoc, over-long root
  *.mill lines wrapped
