+++ kb/bundles/intent/index.md
@@ -20,7 +20,7 @@ produced, which is what keeps the two in step.
 
 _Generated by `kb refresh` — do not edit below the marker. Last built 2026-08-22._
 
-## In progress (11)
+## In progress (10)
 
 * [0003 Capability coverage for morphir-scala — docs](/0003-capability-coverage-for-morphir-scala.md) - Describe what morphir-scala does today, so released intent has capabilities to link to.
 * [0008 Buildkit core task graph — feature](/0008-buildkit-core-task-graph.md) - Provide a Morphir-agnostic, Kyo-native task graph for composing and executing typed cross-platform pipelines.
@@ -32,7 +32,6 @@ _Generated by `kb refresh` — do not edit below the marker. Last built 2026-08-
 * [0030 morphir-desktop Electron app — feature](/0030-morphir-desktop-electron-app.md) - Build morphir/desktop, an unpublished Electron app hosting morphir-ui with Scala.js in both processes over a kyo-jsonrpc seam.
 * [0031 Publish the morphir-desktop application — feature](/0031-publish-the-morphir-desktop-application.md) - Ship the Electron desktop app from CI to GitHub Releases and Maven Central, with per-asset checksums and Mill-style platform naming.
 * [0033 Markdown compilation — feature](/0033-markdown-compilation.md) - Compile the Markdown AST through one fold algebra to two cross-platform writers: kyo-ui for the browser and ScalaTags for CommonMark conformance.
-* [0035 GitHub Flavored Markdown profile — breaking, feature](/0035-github-flavored-markdown-profile.md) - Recognize the five GitHub Flavored Markdown extensions behind an opt-in profile, measured against the GFM specification's own example set.
 
 ## In refinement (2)
 
@@ -58,13 +57,14 @@ _Generated by `kb refresh` — do not edit below the marker. Last built 2026-08-
 * [0028 Remove or absorb morphir/toolkit — removal](/0028-remove-or-absorb-morphir-toolkit.md) - Delete or absorb the unwired morphir/toolkit directory so it is not revived as a published-library namespace.
 * [0034 Project-owned kb templates and conformance scoring — docs](/0034-project-owned-kb-templates-and-conformance-scoring.md) - Let each repository carry its own document templates and its own bar for a good document, instead of hardcoding Morphir's in the kb skill.
 
-## Released (5)
+## Released (6)
 
 * [0001 Intent management in the knowledge base — feature](/0001-intent-management-in-the-knowledge-base.md) - Record features, enhancements and bugs as prose in kb/, with a lifecycle and a SQLite index.
 * [0002 Release labels for performance, security and deprecation — build](/0002-release-labels-for-performance-security-and-deprecation.md) - Add the three release-drafter labels the intent kind vocabulary needs.
 * [0005 Enforce knowledge base checks in CI — build](/0005-enforce-knowledge-base-checks-in-ci.md) - Run kb check and intent check on every pull request, so the obligations bite instead of being advisory.
 * [0006 Test suite for the kb and intent skill — test](/0006-test-suite-for-the-kb-and-intent-skill.md) - Cover the skill's parsing, scaffolding, checks and index with an executable suite that runs in CI.
 * [0032 Independent version streams — feature](/0032-independent-version-streams.md) - Version the libraries, the Mill plugin family and the desktop application from their own changelogs and tag streams.
+* [0035 GitHub Flavored Markdown profile — breaking, feature](/0035-github-flavored-markdown-profile.md) - Recognize the five GitHub Flavored Markdown extensions behind an opt-in profile, measured against the GFM specification's own example set.
 
 ## Cancelled (1)
